feat: fully match current map-interpreter behavior
