+++ src/core/profile-provider/bound-profile-provider.ts
@@ -14,7 +14,7 @@ import type {
   MapInterpreterError,
   ProfileParameterError,
 } from '../../interfaces';
-import type { NonPrimitive, Result, UnexpectedError } from '../../lib';
+import type { NonPrimitive, Result, SDKExecutionError, UnexpectedError } from '../../lib';
 import {
   castToNonPrimitive,
   err,
@@ -46,7 +46,7 @@ export interface IBoundProfileProvider {
   ): Promise<
     Result<
       TResult,
-      ProfileParameterError | MapInterpreterError | UnexpectedError
+      ProfileParameterError | MapInterpreterError | UnexpectedError | SDKExecutionError
     >
   >;
 }
@@ -102,7 +102,7 @@ export class BoundProfileProvider implements IBoundProfileProvider {
   ): Promise<
     Result<
       TResult,
-      ProfileParameterError | MapInterpreterError | UnexpectedError
+      ProfileParameterError | MapInterpreterError | UnexpectedError | SDKExecutionError
     >
   > {
     this.fetchInstance.metadata = {
