+++ src/interfaces/errors/usecase.errors.ts
@@ -1,5 +1,6 @@
+import type { SDKExecutionError } from '../../lib';
 import type { MapInterpreterError } from './map-interpreter.errors';
 import type { ProfileParameterError } from './profile-parameter-validator.errors';
 
 // TODO
-export type PerformError = ProfileParameterError | MapInterpreterError;
+export type PerformError = ProfileParameterError | MapInterpreterError | SDKExecutionError;
